Refactoring. Other changes.
