+++ Minio.Examples/Cases/SetObjectRetention.cs
@@ -15,6 +15,7 @@
  */
 using System;
 using System.Threading.Tasks;
+using Minio.DataModel.ObjectLock;
 
 namespace Minio.Examples.Cases
 {
