+++ Minio/DataModel/ObjectLock/DefaultRetention.cs
@@ -13,37 +13,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 using System;
-using System.Xml;
 using System.Xml.Serialization;
 
-namespace Minio
+namespace Minio.DataModel.ObjectLock
 {
-    public enum RetentionMode
-    {
-        GOVERNANCE,
-        COMPLIANCE
-    }
     [Serializable]
-    [XmlRoot(ElementName = "Retention", Namespace = "http://s3.amazonaws.com/doc/2006-03-01/")]
-    public class ObjectRetentionConfiguration
+    [XmlRoot(ElementName = "DefaultRetention", Namespace = "http://s3.amazonaws.com/doc/2006-03-01/")]
+    public class DefaultRetention
     {
+        public DefaultRetention()
+        {
+        }
+
+        public DefaultRetention(int days, RetentionMode mode)
+        {
+            this.Days = days;
+            this.Mode = mode;
+        }
+        [XmlElement("Days")]
+        public int Days { get; set; }
         [XmlElement("Mode")]
         public RetentionMode Mode { get; set; }
 
-        [XmlElement("RetainUntilDate")]
-        public string RetainUntilDate { get; set; }
-
-        public ObjectRetentionConfiguration()
-        {
-            this.RetainUntilDate = null;
-        }
-
-        public ObjectRetentionConfiguration(DateTime date, RetentionMode mode = RetentionMode.GOVERNANCE)
-        {
-            this.RetainUntilDate = utils.To8601String(date);
-            this.Mode = mode;
-        }
     }
 }
